+++ Gateway/Request/EmbedForm/CaptureRequest.php
@@ -136,8 +136,8 @@ class CaptureRequest implements BuilderInterface
                 $charge['payment_method']['token_id'] = $token;
                 break;
             
-            case ConfigProvider::PAYMENT_METHOD_OXXO:
-            case ConfigProvider::PAYMENT_METHOD_SPEI:
+            case ConfigProvider::PAYMENT_METHOD_CASH:
+            case ConfigProvider::PAYMENT_METHOD_BANK_TRANSFER:
                 $reference = $payment->getAdditionalInformation('reference');
                 $expireAt = $this->_conektaHelper->getExpiredAt();
                 $charge['payment_method']['reference'] = $reference;