+++ Gateway/Response/EmbedForm/TxnIdHandler.php
@@ -51,9 +51,8 @@ class TxnIdHandler implements HandlerInterface
             case ConfigProvider::PAYMENT_METHOD_CREDIT_CARD:
                 $this->handleCreditCard($payment, $response);
                 break;
-            
-            case ConfigProvider::PAYMENT_METHOD_OXXO:
-            case ConfigProvider::PAYMENT_METHOD_SPEI:
+            case ConfigProvider::PAYMENT_METHOD_CASH:
+            case ConfigProvider::PAYMENT_METHOD_BANK_TRANSFER:
                 $this->handleOffline($payment, $response);
                 break;
         }