+++ Helper/Data.php
@@ -319,23 +319,23 @@ class Data extends Util
     }
 
     /**
-     * Is Oxxo enabled
+     * Is cash enabled
      *
      * @return bool
      */
-    public function isOxxoEnabled()
+    public function isCashEnabled()
     {
-        return (boolean)$this->getConfigData('conekta_oxxo', 'active');
+        return (boolean)$this->getConfigData('conekta_cash', 'active');
     }
 
     /**
-     * Is Spei enabled
+     * Is Bank Transfer enabled
      *
      * @return bool
      */
-    public function isSpeiEnabled()
+    public function isBankTransferEnabled()
     {
-        return (boolean)$this->getConfigData('conekta_spei', 'active');
+        return (boolean)$this->getConfigData('conekta_bank_transfer', 'active');
     }
 
     /**
@@ -588,7 +588,7 @@ class Data extends Util
         $urlWebhook = $this->getConfigData('conekta/conekta_global', 'conekta_webhook');
         if (empty($urlWebhook)) {
             $baseUrl = $this->_storeManager->getStore()->getBaseUrl();
-            $urlWebhook = $baseUrl . "conekta/webhook/listener";
+            $urlWebhook = $baseUrl . "conekta/webhook/index";
         }
         return $urlWebhook;
     }