+++ view/frontend/web/js/view/payment/method-renderer/embedform.js
@@ -289,14 +289,22 @@ define(
                         theme: 'default'
                     },
                     onCreateTokenSucceeded: function (token) {
+
                     },
                     onCreateTokenError: function (error) {
                         console.error(error);
                     },
                     onFinalizePayment: function (event) {
+                        if (event.charge.paymentMethod.type != 'card' && event.charge.payment_method.type != 'card') {
+                            event.charge.status = 'pending_payment'
+                            event.status = 'pending_payment'
+                        }
                         self.iframOrderData(event);
                         self.beforePlaceOrder();
-                    }
+                    },
+                    onErrorPayment: function(a) {
+                        self.conektaError("Ocurrio un error al procesar el pago. Por favor intente de nuevo");
+                    },
                 });
 
                 $('#conektaIframeContainer').find('iframe').attr('data-cy', 'the-frame');
@@ -403,7 +411,6 @@ define(
             getMinimumAmountMonthlyInstallments: function () {
                 return this.getMethodConfig().minimum_amount_monthly_installments;
             },
-
             validateCheckoutSession: function () {
                 const lifeTime = parseInt(this.getMethodConfig().sessionExpirationTime)
                 const timeToExpire = (lifeTime - 5) * 1000