+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 4.2.0 - 2023/05/18 19:43
+* Feat:
+  - [CHLP-1284] Change name for payment methods OXXO and SPEI to cash and bank transfer. @elvisheredia [#40](https://github.com/conekta/customer-magento-plugin/pull/40)
+    - Set order status to Payment Review when payment method was cash or bank transfer
+    - Show additional info in one success page showing payment reference or CLABE
+  - [CHLP-1284] Change default webhook to resolve error with change status to Cancel or Processing when offline payment was processed @elvisheredia [#40](https://github.com/conekta/customer-magento-plugin/pull/40)
+
 ### 4.1.9 - 2023/03/15 16:34
 [Full Changelog](https://github.com/conekta/customer-magento-plugin/compare/4.1.8...4.1.9)
 * Feat: